move integration etherscan tests to tests/ dir
