+++ ethers-etherscan/src/verify.rs
@@ -153,55 +153,3 @@ impl Client {
         self.post_form(&body).await
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{tests::run_at_least_duration, Client};
-    use ethers_core::types::Chain;
-    use ethers_solc::{Project, ProjectPathsConfig};
-    use serial_test::serial;
-    use std::{path::PathBuf, time::Duration};
-
-    #[allow(unused)]
-    fn init_tracing() {
-        tracing_subscriber::fmt()
-            .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
-            .init();
-    }
-
-    #[tokio::test]
-    #[serial]
-    #[ignore]
-    async fn can_flatten_and_verify_contract() {
-        init_tracing();
-        run_at_least_duration(Duration::from_millis(250), async {
-            let root = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("resources");
-            let paths = ProjectPathsConfig::builder()
-                .sources(&root)
-                .build()
-                .expect("failed to resolve project paths");
-            let project = Project::builder()
-                .paths(paths)
-                .build()
-                .expect("failed to build the project");
-
-            let address = "0x9e744c9115b74834c0f33f4097f40c02a9ac5c33".parse().unwrap();
-            let compiler_version = "v0.5.17+commit.d19bba13";
-            let constructor_args = "0x000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000005f5e1000000000000000000000000000000000000000000000000000000000000000007596179537761700000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000035941590000000000000000000000000000000000000000000000000000000000";
-            let contract = project.flatten(&root.join("UniswapExchange.sol")).expect("failed to flatten contract");
-            let contract_name = "UniswapExchange".to_owned();
-
-            let client = Client::new_from_env(Chain::Mainnet).unwrap();
-
-            let contract =
-                VerifyContract::new(address, contract_name, contract, compiler_version.to_string())
-                    .constructor_arguments(Some(constructor_args))
-                    .optimization(true)
-                    .runs(200);
-            let resp = client.submit_contract_verification(&contract).await.expect("failed to send the request");
-            assert_ne!(resp.result, "Error!"); // `Error!` result means that request was malformatted
-        })
-        .await
-    }
-}
